move core_ext to lib/pru
