+++ lib/pru.rb
@@ -1,4 +1,4 @@
-require 'core_ext'
+require 'pru/core_ext'
 
 class Pru
   VERSION = File.read( File.join(File.dirname(__FILE__),'..','VERSION') ).strip
